pyproject.toml
[tool.poetry]
name = "Quotes"
version = "1.0"
description = "Quotes API is a FastAPI service for generating quote images using HTML templates"
authors = ["Fl1yd <projects@fl1yd.su>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
pillow = "9.5.0"
uvicorn = "^0.30.1"
imgkit = "^1.2.3"
pydantic = "^2.8.2"
dacite = "^1.8.1"
loguru = "^0.7.2"
toml = "^0.10.2"
jinja2 = "^3.1.4"
regex = "^2024.9.11"


[tool.poetry.group.dev.dependencies]
pyrogram = "^2.0.106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
